requirements.txt
fastapi~=0.111.0
python-dotenv~=1.0.1
PyJWT~=2.8.0
uvicorn~=0.30.1
pytest
pytest-cov
slack_sdk